feat: add map rendering to the page
