+++ package.json
@@ -6,6 +6,8 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "leaflet": "^1.9.4",
+    "react-leaflet": "^4.2.1",
     "react-router-dom": "^6.20.0",
     "react-scripts": "5.0.1",
     "sass": "^1.69.5",
