+++ public/index.html
@@ -22,8 +22,10 @@
   <noscript>You need to enable JavaScript to run this app.</noscript>
   <div id="slider-hook"></div>
   <div id="overlay-hook"></div>
-  <div id="map"></div>
-  <div id="root"></div>
+
+  <div id="root">
+    <div id="map"></div>
+  </div>
 
 
 </body>
