+++ src/map/pages/Map.js
@@ -1,23 +1,25 @@
-import React, { useEffect, useState } from 'react';
+// import React, { useEffect, useState } from 'react';
 import ShowMap from '../components/loadMap';
+// import { createPortal } from 'react-dom';
 
 const Map = props => {
-	const [currLocation, setCurrLocation] = useState({});
+	// const [currLocation, setCurrLocation] = useState({});
 
-	useEffect(() => {
-		getLocation();
-	}, []);
+	// useEffect(() => {
+	// 	getLocation();
+	// }, []);
 
-	const getLocation = function () {
-		if (navigator.geolocation)
-			navigator.geolocation.getCurrentPosition(
-				position => setCurrLocation(position.coords),
-				function () {
-					alert('Could not get your position. Try again!');
-				}
-			);
-	};
+	// const getLocation = function () {
+	// 	if (navigator.geolocation)
+	// 		navigator.geolocation.getCurrentPosition(
+	// 			position => setCurrLocation(position.coords),
+	// 			function () {
+	// 				alert('Could not get your position. Try again!');
+	// 			}
+	// 		);
+	// };
 
-	return <ShowMap latitude={currLocation.latitude} longitude={currLocation.longitude} />;
+	return <ShowMap />;
+	// return createPortal(<ShowMap />, document.getElementById('map'));
 };
 export default Map;
